vpudec: Enable VPU secure mode
In secure mode, the VPU expects the virtual address of the shared memory and
the physical address of the secure memory. Retrieve the physisal address of
the secure ION buffer and provide it to the VPU wrapper API.

When the GstBuffer does not have a ION secure memory, provide NULL as the
physical address. The VPU wrapper will use its internal ring buffer to send
the data to the VPU.

Signed-off-by: Alexandre Jutras <alexandre.jutras@linaro.org>